refactor: server のエラー関連を error モジュールに集約
server 直下の handle-error と auth/ 配下の auth-error が別々に置かれ、
エラー変換の責務が散在していたため src/server/error/ にまとめる。

- handle-error.ts / .test.ts を src/server/error/ へ移動
- auth-error.ts / .test.ts を src/server/auth/ から src/server/error/ へ移動
- 各ハンドラの import を @/server/error/* に更新（auth-error は相対 ../auth-error
  から絶対 @/ パスに揃え、import 規約にも沿う形にする）
