+++ application/apps/web/src/server/article/handler/get-articles.ts
@@ -11,7 +11,7 @@ import {
 import { z } from 'zod'
 import CONTEXT_KEY from '@/middleware/context'
 import type { ZodValidatedQueryContext } from '@/middleware/zod-validator'
-import { handleError } from '@/server/handle-error'
+import { handleError } from '@/server/error/handle-error'
 import { type ArticleResponse, toArticleResponse } from '../article-response'
 
 const readStatusEnum = z.enum(['0', '1'])
