+++ application/apps/web/src/server/article/handler/read-article.ts
@@ -4,7 +4,7 @@ import { DIARY_DAYS } from '@trend-diary/domain/article/diary'
 import { z } from 'zod'
 import CONTEXT_KEY from '@/middleware/context'
 import type { ZodValidatedParamJsonContext } from '@/middleware/zod-validator'
-import { handleError } from '@/server/handle-error'
+import { handleError } from '@/server/error/handle-error'
 
 const MS_PER_MINUTE = 60 * 1000
 const MS_PER_DAY = 24 * 60 * MS_PER_MINUTE
