+++ application/apps/web/src/server/auth/handler/login.ts
@@ -3,8 +3,8 @@ import getRdbClient from '@trend-diary/datastore/rdb'
 import { type AuthInput, createAccountUseCase } from '@trend-diary/domain/account'
 import CONTEXT_KEY from '@/middleware/context'
 import type { ZodValidatedContext } from '@/middleware/zod-validator'
-import { handleError } from '@/server/handle-error'
-import toAuthError from '../auth-error'
+import toAuthError from '@/server/error/auth-error'
+import { handleError } from '@/server/error/handle-error'
 import { verifyTurnstile } from '../captcha'
 
 export default async function login(c: ZodValidatedContext<AuthInput>) {
