+++ application/apps/web/src/server/auth/handler/logout.ts
@@ -1,8 +1,8 @@
 import { authClientConfig, PasswordAuthClient } from '@trend-diary/authentication'
 import type { Context } from 'hono'
 import CONTEXT_KEY from '@/middleware/context'
-import { handleError } from '@/server/handle-error'
-import toAuthError from '../auth-error'
+import toAuthError from '@/server/error/auth-error'
+import { handleError } from '@/server/error/handle-error'
 
 export default async function logout(c: Context) {
   const logger = c.get(CONTEXT_KEY.APP_LOG)
