+++ application/apps/web/src/server/auth/handler/passkey-login-start.ts
@@ -2,8 +2,8 @@ import { authClientConfig, PasskeyClient } from '@trend-diary/authentication'
 import type { Context } from 'hono'
 import type { Env } from '@/env'
 import CONTEXT_KEY from '@/middleware/context'
-import { handleError } from '@/server/handle-error'
-import toAuthError from '../auth-error'
+import toAuthError from '@/server/error/auth-error'
+import { handleError } from '@/server/error/handle-error'
 
 export default async function passkeyLoginStart(c: Context<Env>) {
   const logger = c.get(CONTEXT_KEY.APP_LOG)
