+++ application/apps/web/src/server/error/auth-error.test.ts
