+++ application/apps/web/src/server/error/auth-error.ts
