+++ application/apps/web/src/server/error/handle-error.ts
